Improved design of split-brain SPI
* moved SplitBrainEntryViews to an internal package, since it uses
  internals of HZ and shouldn't be in a public SPI package
* moved SplitBrainMergePolicyDataSerializerHook to an internal package,
  since all other serializer hooks are internal as well
* made SimpleSplitBrainEntryView a first class citizen, since it will
  be needed by users who implement their own split-brain capable service
