+++ hazelcast/src/main/java/com/hazelcast/cardinality/impl/operations/MergeOperation.java
@@ -26,7 +26,7 @@ import com.hazelcast.spi.SplitBrainMergePolicy;
 import java.io.IOException;
 
 import static com.hazelcast.cardinality.impl.CardinalityEstimatorDataSerializerHook.MERGE;
-import static com.hazelcast.spi.merge.SplitBrainEntryViews.createSplitBrainMergeEntryView;
+import static com.hazelcast.spi.impl.merge.SplitBrainEntryViews.createSplitBrainMergeEntryView;
 
 /**
  * Contains a mergeable {@link HyperLogLog} instance for split-brain healing with a {@link SplitBrainMergePolicy}.
