+++ hazelcast/src/main/resources/META-INF/services/com.hazelcast.DataSerializerHook
@@ -32,4 +32,4 @@ com.hazelcast.projection.impl.ProjectionDataSerializerHook
 com.hazelcast.config.ConfigDataSerializerHook
 com.hazelcast.journal.EventJournalDataSerializerHook
 com.hazelcast.flakeidgen.impl.FlakeIdGeneratorDataSerializerHook
-com.hazelcast.spi.merge.SplitBrainMergePolicyDataSerializerHook
+com.hazelcast.spi.impl.merge.SplitBrainMergePolicyDataSerializerHook
