+++ hazelcast/src/test/java/com/hazelcast/wan/impl/WanReplicationTest.java
@@ -40,8 +40,8 @@ import com.hazelcast.spi.OperationFactory;
 import com.hazelcast.spi.SplitBrainMergeEntryView;
 import com.hazelcast.spi.SplitBrainMergePolicy;
 import com.hazelcast.spi.impl.NodeEngineImpl;
+import com.hazelcast.spi.impl.merge.SplitBrainEntryViews;
 import com.hazelcast.spi.impl.operationservice.InternalOperationService;
-import com.hazelcast.spi.merge.SplitBrainEntryViews;
 import com.hazelcast.spi.serialization.SerializationService;
 import com.hazelcast.test.AssertTask;
 import com.hazelcast.test.HazelcastParallelClassRunner;
